Adding Emitter
Emiter object will connect through a COM port and send bytes to the
robot. Interpreter will open the connection and pass the handle to
Emitter. So interpreter can run the code in the simulator, while the
real robot follows.
The issue is how to receive date, since ED7220C talks via events, and
we cant handle serial events with Coppelia Serial API.
The Emitter will send a predefined sequence of bytes with ASCII
commands, for each KEYWORD registered. We use a switch function to
register keywords and the respective function to execute, instead of
a big 'IF ELIF ELSE' statement.