Module for Switch Functionality
Lua has no switch statement. Using IF ELSEIF ELSE is fine, but not
elegant. This module adds switch like functionality, using tables
and function as objects.
There is a Object Oriented version, wich creates a Switch object,
that constructs a table with the 'build():case()..' method. Then
a 'switch(value,...)' method executes the function associated with
the key. This is better for long tables, as it is constructed once.
The second version is a switch() function, wich constructs the table
and excecute in one go. This is better for switchs that execute only
once.
Both are milliseconds slower that 'IF ELSEIF' statements, but makes
the code a little bit more functional.