Files for building 'libserial' Lua dynamic library
These files are used to compile the libserial libray for the
Standalone Interpreter. It compiles with MSVC and nmake. It
did not work with MYSYS GNU.
Must compile with C++20 minimum as it uses std::format.

[include] - Include directory for Lua headers.
[libserial.cpp] - Lua library, that wraps SerialPort C++ libray.
Built with Lua 5.4.
[SerialPort.h] - SerialPort class header.
[SerialPort.cpp] - SerialPort class library built upon Windows API.
Currently working only on Synchronous mode (no overlapped).
[CMakeLists.txt] - Cmake Configuration File. It generates a 'build'
directory, where the final libserial.dll is.