Added new Symbols
Added HARDHOME keyword, POW as power operation for operator '^',
and added floating point indentifiers, preceded by the symbol '@'.

Robotalk uses @ to denote variables that holds floats, but in Lua,
every number can be coerced to float. So assigning 3.14 to 'R',
gives no errors, neither 2 to @R. It is just to follow robotalk
rules. But the benefit is that now there is the double of variables.