Implemented Remaining Instructions to Emitter
Added Runtime namespace to manage runtime events and errors.
Added power operation to Evaluator.

The COM port is now the second command line argument, and the
port is set before the Interpreter is created.

Added Register function to Emitter using the Switch object,
so functions are declared once.

Implemented remaining functions to handle the Robot Instructions