Implemented Fast Mode. Changed CRLF to LF
The Controller runs on Coppelia Threaded enviroment. Lua threads
are not real. They are implemented on top of Lua coroutines. To
simulate threads, a scheduler preemptively suspends and resumes
registered function. But this adds a lot latency, so the controller
becomes really slow for fast operations or math computations.

Fast mode disables auto yield when running the interpreter, and
enables it back when running 'robot' or 'emitter'. This allows 20x
times faster execution for some operations, such as TYPE and FOR loops.
Code that contains PAUSE does not experience any diference.

Fast Mode has the disavantage of blocking UI, so a warning is given
to the user.

Windows adds CR or '\r' at the end of lines, but commits are made
from WSL, so git complains a lot. Removed '\r' from all files.